Add preset for Links Addon
