+++ addons/links/src/preset/addDecorator.ts
@@ -1,0 +1,3 @@
+import { withLinks } from '../index';
+
+export const decorators = [withLinks];
